backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
pytest==8.2.2
